refactor: move borsh utils to own file
